Improve error handling and move endpoints to ts
