+++ app/models/server/raw/LivechatVisitors.ts
@@ -1,10 +1,26 @@
 import { escapeRegExp } from '@rocket.chat/string-helpers';
-import { AggregationCursor, Cursor, FilterQuery, FindOneOptions } from 'mongodb';
+import { AggregationCursor, Cursor, FilterQuery, FindOneOptions, WithoutProjection } from 'mongodb';
 
 import { BaseRaw } from './BaseRaw';
 import { ILivechatVisitor } from '../../../../definition/ILivechatVisitor';
 
 export class LivechatVisitorsRaw extends BaseRaw<ILivechatVisitor> {
+	findOneById(_id: string, options: WithoutProjection<FindOneOptions<ILivechatVisitor>>): Promise<ILivechatVisitor | null> {
+		const query = {
+			_id,
+		};
+
+		return this.findOne(query, options);
+	}
+
+	getVisitorByToken(token: string, options: WithoutProjection<FindOneOptions<ILivechatVisitor>>): Promise<ILivechatVisitor | null> {
+		const query = {
+			token,
+		};
+
+		return this.findOne(query, options);
+	}
+
 	getVisitorsBetweenDate({ start, end, department }: { start: Date; end: Date; department: string }): Cursor<ILivechatVisitor> {
 		const query = {
 			_updatedAt: {
