+++ definition/ILivechatVisitor.ts
@@ -33,3 +33,21 @@ export interface ILivechatVisitor extends IRocketChatRecord {
 	host?: string;
 	visitorEmails?: IVisitorEmail[];
 }
+
+export interface ILivechatVisitorDTO {
+	id: string;
+	token: string;
+	name: string;
+	email: string;
+	department: string;
+	phone: string | { number: string };
+	username: string;
+	customFields: {
+		key: string;
+		value: string;
+		overwrite: boolean;
+	}[];
+	connectionData: {
+		httpHeaders: Record<string, string>;
+	};
+}
